requirements.txt
pandas
numpy
streamlit
python-dotenv
groq
matplotlib
seaborn
openpyxl